+++ fitConnect/src/main/java/com/example/fitconnect/controller/chat/ChatRoomController.java
@@ -5,9 +5,9 @@ import com.example.fitconnect.config.service.CommonService;
 import com.example.fitconnect.domain.chat.domain.ChatRoom;
 import com.example.fitconnect.domain.chat.dto.ChatRoomRegistrationDto;
 import com.example.fitconnect.domain.chat.dto.ChatRoomUpdateDto;
-import com.example.fitconnect.service.chat.ChatRoomCreationService;
-import com.example.fitconnect.service.chat.ChatRoomDeleteService;
-import com.example.fitconnect.service.chat.ChatRoomUpdateService;
+import com.example.fitconnect.service.chat.chatRoom.ChatRoomCreationService;
+import com.example.fitconnect.service.chat.chatRoom.ChatRoomDeleteService;
+import com.example.fitconnect.service.chat.chatRoom.ChatRoomUpdateService;
 import jakarta.servlet.http.HttpSession;
 import jdk.jshell.Snippet.Status;
 import lombok.RequiredArgsConstructor;
