+++ fitConnect/src/main/java/com/example/fitconnect/service/chat/chatMessage/ChatMessageCreationService.java
@@ -7,8 +7,7 @@ import com.example.fitconnect.domain.chat.domain.ChatRoom;
 import com.example.fitconnect.domain.chat.dto.ChatMessageRegistrationDto;
 import com.example.fitconnect.domain.user.domain.User;
 import com.example.fitconnect.repository.chat.ChatMessageRepository;
-import com.example.fitconnect.repository.chat.ChatRoomRepository;
-import com.example.fitconnect.service.chat.ChatRoomFindService;
+import com.example.fitconnect.service.chat.chatRoom.ChatRoomFindService;
 import com.example.fitconnect.service.user.UserFindService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
