+++ fitConnect/src/main/java/com/example/fitconnect/service/chat/chatRoom/ChatRoomDeleteService.java
@@ -1,30 +1,30 @@
-package com.example.fitconnect.service.chat;
+package com.example.fitconnect.service.chat.chatRoom;
 
 import com.example.fitconnect.config.error.ErrorMessages;
 import com.example.fitconnect.config.exception.EntityNotFoundException;
 import com.example.fitconnect.domain.chat.domain.ChatRoom;
-import com.example.fitconnect.domain.chat.dto.ChatRoomUpdateDto;
 import com.example.fitconnect.repository.chat.ChatRoomRepository;
-import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class ChatRoomUpdateService {
-
-    private final ChatRoomRepository chatRoomRepository;
+public class ChatRoomDeleteService {
 
     private final ChatRoomFindService chatRoomFindService;
 
-    @Transactional
-    public void updateTitle(ChatRoomUpdateDto chatRoomUpdateDto,Long userId) {
+    private final ChatRoomRepository chatRoomRepository;
 
-        ChatRoom chatRoom = chatRoomFindService.findCharRoomByChatRoomId(
-                        chatRoomUpdateDto.getChatRoomId())
+    public void deleteChatRoom(Long userId, Long chatRoomId) {
+
+
+        ChatRoom chatRoom = chatRoomFindService.findCharRoomByChatRoomId(chatRoomId)
                 .orElseThrow(() -> new EntityNotFoundException(ErrorMessages.CHATROOM_NOT_FOUND));
 
         chatRoom.validateCreator(userId);
-        chatRoom.update(chatRoomUpdateDto.getTitle());
+        chatRoomRepository.delete(chatRoom);
+
     }
+
 }
+
